+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [0.4.3] - 2026-04-07
+
+#### Added
+
+- Cap tick density for readability by @pecastro in #19
+
 
 ## [0.4.2] - 2026-04-03